Up to revno 12398, i.e. the R3.14.12.3 release.

+++ src/libCom/test/cvtFastPerform.cpp
@@ -60,9 +60,9 @@ void Test :: execute ()
     
     for ( unsigned i = 0; i < 3; i++ ) {
         double mVal = rand ();
-        mVal /= ((double) RAND_MAX + 1);
+        mVal /= (RAND_MAX + 1.0);
         double fEVal = rand ();
-        fEVal /= ((double) RAND_MAX + 1);
+        fEVal /= (RAND_MAX + 1.0);
         fEVal *= DBL_MAX_EXP - DBL_MIN_EXP;
         fEVal += DBL_MIN_EXP;
         int eVal = static_cast < int > ( fEVal + 0.5 );
@@ -72,7 +72,7 @@ void Test :: execute ()
             _measure ();
         }
         _srcVal = rand ();
-        _srcVal /= ((double) RAND_MAX + 1);
+        _srcVal /= (RAND_MAX + 1.0);
         _srcVal *= 10.0;
         _srcVal -= 5.0;
         for ( _prec = lowPrecision;
